Add files via upload
Here are the instructions I had to follow for writing this java code. Its a simple-ish app that is used to grade students using weighted averages that writes eveything to a .txt file. .....

Part | - Scenario

The instructors at the community college are happy with the weighted grade averaging program you wrote for them and want to expand it. You have been asked by the community college to modify the existing program to include additional functionality for the faculty.

The new program will accept data from the instructor using a menu which will ask the instructor which category of grade they want to enter. Once the instructor enters a category, they will then be able to enter the category grade. The program will continue until the instructor indicates they no longer want to enter grades.

Once the instructor has stopped entering grades, the program will calculate the weighted average, determine the letter grade, decide if the student is passing and write the data to a file. The program will also read the data from the file and display the information on the screen.

Part II - Create a new class from an existing class.

    ﻿﻿﻿Download the file that was uploaded to Blackboard for the Project 1 program.
    ﻿﻿﻿Unzip the file.
    ﻿﻿﻿Create a new folder with your LastNameFirstName_Project as the folder name.
    ﻿﻿﻿Copy the file named Projecti.java to the new LastNameFirstName_Project folder.
    ﻿﻿﻿Open the file in the new folder and rename the class with the name Project2.
    ﻿﻿﻿Use "Save As" to save the file as Project2.java.
    ﻿﻿﻿Use this file to continue with the project and modify the program.

Part III - Add additional input to the program.

    ﻿﻿Display a set of choices at a prompt that asks the instructor to select a category of grade they want to enter.
    ﻿﻿Validate the input, if the user enters a value that is not on the list, ask them to re-enter a value until a valid category value is entered. (Hint: you can use a loop for this.)
    ﻿﻿Helpful hint: Display values to the instructor and ask that the user enter values that are easy to compare. For example, 1-Training, 2-Lab, etc. is easy to compare because numeric values can use the "==" relational operator. The program can also request characters such as t-Training, I-Lab, etc. Asking a user to enter an entire word can be tricky if they misspell the word.

Part IV - Processing

    ﻿﻿Once the instructor enters a category, ask the instructor to enter that category's weighted grade value (this code is already in the program).
    ﻿﻿Continue to prompt the instructor to enter a category until the instructor no longer wants to enter a grade category. (Hint: use a loop with a sentinel value such as 'y for yes to continue the entry and have the loop stop if the user enters the sentinel value such as 'n' to stop.)
    ﻿﻿When the instructor no longer wants to continue entering category grades, then the program should:
    ﻿﻿calculate the final grade (this is already in the program).
    ﻿﻿determine the letter grade using a 10-point scale.
    ﻿﻿determine if the student is passing when the student's grade exceeds the passing rate of 60% or greater.


Part V - Output to file and screen.

File output - Writing data to a file.

    ﻿﻿Write the same data that was displayed to the screen in Project 1 (with the addition of the newly added data from this project) to a file named "Project2Grades.txt".
    ﻿﻿Make sure to append the data instead of overwriting the data so there is a record for each student in the file. (Hint: use the FileWriter class to create an object and pass it to the PrintWriter class constructor along with the "true" Boolean value. See the FileWriterDemo_modified example in Blackboard for help on this).
    ﻿﻿Make sure the program closes the file.

The student grade report in the file should contain:

    ﻿﻿Report title (first line)
    ﻿﻿Student's first and last name on the second line (i.e., John Smith)
    ﻿﻿Student ID on the third line
    ﻿﻿Display the four category averages entered by the instructor (four separate lines, one for each category) along with a label that that will describe which category is being displayed.(Hint: use concatenation to do this).
    ﻿﻿Display the result of the weighted average calculation of the four grade categories.
    ﻿﻿Display the letter grade.
    ﻿﻿Display a message indicating whether the student passed the course.

Screen output - Displaying data to a file.

o. Once the instructor has finished entering the data, and the data has been written to the file, then read the data from the file named "Project2Grades.txt" and display all the lines of data in the file to the screen.

Make sure the program closes the file.