Add files via upload
The FitnessClubApp file is the main application that allows users to input, save, and display member information, writing data to a file and retrieving it as needed. The Member file defines the structure of a member with attributes like name, age, phone, and email, while also keeping track of the total number of members created.